move fixtures so they can be reused outside dataset module
