+++ qcodes/tests/test_plot_utils.py
@@ -2,23 +2,18 @@
 Tests for `qcodes.utils.plotting`.
 """
 import matplotlib
+
 # set matplotlib backend before importing pyplot
 matplotlib.use("Agg")
 
-from pytest import fixture
 from matplotlib import pyplot as plt
+from pytest import fixture
 
-# we only need `dataset` here, but pytest does not discover the dependencies
-# by itself so we also need to import all the fixtures this one is dependent
-# on.
-# pylint: disable=unused-import
-from qcodes.tests.dataset.conftest import (empty_temp_db,
-                                           experiment, dataset)
-
-from qcodes.tests.common import default_config
-from qcodes.dataset.plotting import plot_by_id
-from .dataset_generators import dataset_with_outliers_generator
 import qcodes
+from qcodes.dataset.plotting import plot_by_id
+from qcodes.tests.common import default_config
+
+from .dataset_generators import dataset_with_outliers_generator
 
 
 @fixture(scope='function')
